creator/src/main/java/net/openhft/chronicle/releasenotes/creator/model/Label.java
package net.openhft.chronicle.releasenotes.creator.model;

import static java.util.Objects.requireNonNull;

import java.util.Objects;

public final class Label {

    private final String name;

    public Label(String name) {
        this.name = requireNonNull(name);
    }

    public String getName() {
        return name;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        final Label label = (Label) o;
        return name.equals(label.name);
    }

    @Override
    public int hashCode() {
        return Objects.hash(name);
    }

    @Override
    public String toString() {
        return "Label{" +
                "name='" + name + '\'' +
                '}';
    }
}
